+++ packages/db/prisma/migrations/20260921212308_add_apikey/migration.sql
@@ -1,0 +1,32 @@
+-- CreateTable
+CREATE TABLE "apikey" (
+    "id" TEXT NOT NULL,
+    "name" TEXT,
+    "start" TEXT,
+    "prefix" TEXT,
+    "key" TEXT NOT NULL,
+    "userId" TEXT NOT NULL,
+    "refillInterval" INTEGER,
+    "refillAmount" INTEGER,
+    "lastRefillAt" TIMESTAMP(3),
+    "enabled" BOOLEAN NOT NULL DEFAULT true,
+    "rateLimitEnabled" BOOLEAN NOT NULL DEFAULT true,
+    "rateLimitTimeWindow" INTEGER,
+    "rateLimitMax" INTEGER,
+    "requestCount" INTEGER NOT NULL DEFAULT 0,
+    "remaining" INTEGER,
+    "lastRequest" TIMESTAMP(3),
+    "expiresAt" TIMESTAMP(3),
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP(3) NOT NULL,
+    "permissions" TEXT,
+    "metadata" TEXT,
+
+    CONSTRAINT "apikey_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE INDEX "apikey_userId_idx" ON "apikey"("userId");
+
+-- AddForeignKey
+ALTER TABLE "apikey" ADD CONSTRAINT "apikey_userId_fkey" FOREIGN KEY ("userId") REFERENCES "user"("id") ON DELETE CASCADE ON UPDATE CASCADE;